Ask before deleting itinerary and tour, Update team button

+++ resources/views/backend/itinerary/index.blade.php
@@ -83,7 +83,16 @@
 <script>
     $('.delete').click(function() {
                 let id = $(this).data('id');      
-                
+                Swal.fire({
+                    title: 'Are you sure?',
+                    text: "You won't be able to revert this!",
+                    type: 'warning',
+                    showCancelButton: true,
+                    confirmButtonColor: '#3085d6',
+                    cancelButtonColor: '#d33',
+                    confirmButtonText: 'Yes, delete it!'
+                }).then((result) => {
+            if (result.value) {
                 $.ajax({
                     
                     type: 'DELETE',
@@ -93,15 +102,16 @@
                         'id': id
                     },
                     success: function (data) {
-                        $('.item' + data['id']).remove();
+                        $('.item' + id).remove();
                         Toast.fire({
                             type: 'success',
                             title: 'Deleted  successfully'
                         })  
                     }
                 });
-            });
-           
+            }
+         })
+    });           
 </script>
 <script src="https://cdn.tiny.cloud/1/8t3cusqbsgxjxtrx0cesy6fo1sdkesg3rsg41aky7y8m430h/tinymce/5/tinymce.min.js"></script>
 <script>
